Add aes256 encryption stream (no chunk for now)
