--- cmd/client/rrrr.sh
@@ -1,9 +1,0 @@
-sudo stty -F /dev/ttyACM0 115200 raw 
-sudo cat /dev/ttyACM0 # >./pitch.txt
-#-s 
-#sudo ./allystar-parser.base /dev/ttyACM0 -s    | nc -lp 10900
-
-
-#./AgroSlalomHardware/src/allystar-parser/build/allystar-parser
-
-
