+++ internal/server/grpcserver/interceptor/interceptor.go
@@ -1,3 +1,4 @@
+// Package interceptor - for grcp call
 package interceptor
 
 import (
